generic manifold tests reusable by other packages
