+++ src/Manifolds.jl
@@ -166,8 +166,18 @@ function __init__()
         using .OrdinaryDiffEq: ODEProblem, AutoVern9, Rodas5, solve
         include("ode.jl")
     end
+
+    @require Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40" begin
+        @require ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210" begin
+            @require ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267" begin
+                include("test.jl")
+            end
+        end
+    end
+
     return nothing
 end
+
 #
 export CoTVector, Manifold, MPoint, TVector, Manifold
 export AbstractSphere
